### 2024-12-16 d510db2
Quote affiliations
