### 2024-12-17 #7590 @danielskatz
👋 @LouiseDck - please add countries to all affiliations.

I'm also unsure why you use "{funkyheatmap}" in your paper in a number of places. Are the {}s coming from some type of formatting instruction?

Also, there is a References heading missing at the end of the paper, so the references look like they are part of the Acknowledgments. Please add this.

Finally, I am going to ask the JOSS editors to take a look at this, as it's not clear to me that this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS. You should hear back in a couple of weeks or so, due to the holidays.
