+++ paper/paper.md
@@ -126,3 +126,4 @@ Y.S. acknowledges support from the Flemish Government under the “Onderzoekspro
 A.C. and Y.S. are funded by an FWO TBM grant (Research Foundation – Flanders, T000119N), and Ghent University Special Research Fund (BOF18-GOA-024).
 W.S. was supported by a Marie Skłodowska-Curie fellowship (101028476).
 
+# References
