### 2024-12-17 #7590 @LouiseDck
Thanks @danielskatz for the speed of your reply! I've adapted the references and added the references section. I hope this will be fine now :crossed_fingers:

The curly-brace thing is a convention in the R community, and is used to denote the name of a package. I don't know where this convention originated, but it is common and still in use, e.g: [here](https://www.sc-best-practices.org/introduction/interoperability.html#nomenclature)

If I may make a case for this piece of software: I think benchmarking is a key part of the scientific process, and software like this enhances the usability of these benchmarks -- For big benchmarks it's a very useful way to communicate results clearly (you simply can't do this effectively with only a huge table full of numbers). For that reason I do think it fits within "extracting knowledge from data" or "supports the execution of research experiments" (as a benchmarking study is a kind of research experiment.
But of course I will await the editorial review, thanks for the consideration!
