### 2024-12-17 72c7f50
Paper update (#52)
* Quote affiliations

* Update affiliations

* Typo

* Add references

* Typo
