+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   executable: executable.html
   mtcars: mtcars.html
   scIB: scIB.html
-last_built: 2024-12-17T07:50Z
+last_built: 2024-12-17T15:52Z
 urls:
   reference: https://funkyheatmap.github.io/funkyheatmap/reference
   article: https://funkyheatmap.github.io/funkyheatmap/articles
