### 2024-12-17 #7590 @rcannood
Thanks for being the managing EiC for this submission!

> Finally, I am going to ask the JOSS editors to take a look at this, as it's not clear to me that this is research software as defined by JOSS. You should hear back in a couple of weeks or so, due to the holidays.

Just to weigh in on this -- I would argue it definitely falls under JOSS' definition of research software:

> JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the **execution of research experiments; extracts knowledge from large data sets;** offers a mathematical library, or similar. While useful for many areas of research, pre-trained machine learning models and notebooks are not in-scope for JOSS.

Since this package was developed and specifically catered to being able to showcase results from (bioinformatics) benchmarking studies. It thus supports the execution of research experiments, and extracts knowledge from the raw benchmarking results. A showcase of which studies (derivatives of) funkyheatmap have been used in is shown [here](https://funkyheatmap.github.io/).

An example of how funkyheatmap extracts knowledge from large datasets: on the openproblems website funkyheatmap is used to summarise the results ( https://openproblems.bio/results/label_projection/ ) based on the raw data from the benchmark ( https://openproblems.bio/results/label_projection/data/results.json ).

Thank you for taking this into consideration and hope for a positive response.
