### 2025-01-06 #7590 @jromanowska
@LouiseDck , @rcannood - I have two suggestions for improving the paper:

- _add a short description of how the visualization is created_ - I'm thinking about something like "To use our tool, the user needs to provide... " and "The best performance is with data being not larger than..."
- reword the first paragraph on page 3 (lines 52-55) - the text doesn't really fit to the previous text (especially the two first sentences)
