### 2025-01-14 #7590 @rhenkin
> 👋 @strengejacke & @rhenkin, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @jromanowska yes happy to help!
