### 2025-01-14 #7590 @jromanowska
Wonderful, @rhenkin ! We'll need to wait for another reviewer to start the process. In the meantime, you can look around in the repo of the submitted software or just wait :)

If you know someone who might be willing to be the second reviewer, please, write their github handles without `@` or their email addresses.
