### 2025-01-20 #7698 @editorialbot
[REVIEW]: funkyheatmap: Visualising data frames with mixed data types
**Submitting author:** @LouiseDck (<a href="http://orcid.org/0000-0001-8100-6823">Louise Deconinck</a>)
**Repository:** https://github.com/funkyheatmap/funkyheatmap
**Branch with paper.md** (empty if default branch):
**Version:** 0.5.2
