### 2025-01-21 funkyheatmap/funkyheatmap#54 @rhenkin
Warning in README example
Hi all,

I'm reviewing the JOSS openjournals/joss-reviews#7698 submission so taking a look at the package now.
I started with the example in the README and there is an issue with the expand argument:
```
Warning message:
In funky_heatmap(data, column_info = column_info, expand = list(xmax = 4)) :
  Argument `expand` is deprecated. Use `position_arguments(expand_* = ...)` instead.
```
The image renders correctly but I get legends too, in the README there are no legends.

--
Rafael
