### 2025-01-23 funkyheatmap/funkyheatmap#54 @rcannood
Thanks Rafael! I created a PR to fix this issue.
