+++ paper/render_figure.R
@@ -33,3 +33,37 @@ df <- tibble::tribble(
   "Proportional data", "80% success, 10% OOM, 10% failed", "pie"
 )
 knitr::kable(df)
+
+library(funkyheatmap)
+library(dplyr, warn.conflicts = FALSE)
+library(tibble, warn.conflicts = FALSE)
+
+data("mtcars")
+
+data <- mtcars %>%
+  rownames_to_column("id") %>%
+  arrange(desc(mpg))
+
+column_info <- tribble(
+  ~id,     ~group,         ~name,                      ~geom,        ~palette,   
+  "id",    "",             "",                         "text",       NA,         
+  "mpg",   "overall",      "Miles / gallon",           "bar",        "palette1",
+  "cyl",   "overall",      "Number of cylinders",      "bar",        "palette2",
+  "disp",  "group1",       "Displacement (cu.in.)",    "funkyrect",  "palette1", 
+  "hp",    "group1",       "Gross horsepower",         "funkyrect",  "palette1", 
+  "drat",  "group1",       "Rear axle ratio",          "funkyrect",  "palette1", 
+  "wt",    "group1",       "Weight (1000 lbs)",        "funkyrect",  "palette1", 
+  "qsec",  "group2",       "1/4 mile time",            "circle",     "palette2", 
+  "vs",    "group2",       "Engine",                   "circle",     "palette2", 
+  "am",    "group2",       "Transmission",             "circle",     "palette2", 
+  "gear",  "group2",       "# Forward gears",          "circle",     "palette2", 
+  "carb",  "group2",       "# Carburetors",            "circle",     "palette2",
+)
+
+g2 <- funky_heatmap(
+  data,
+  column_info = column_info,
+  position_args = position_arguments(expand_xmax = 4)
+)
+
+ggsave("paper/figure2.jpg", g2, width = g2$width, height = g2$height)
