+++ README.Rmd
@@ -78,12 +78,16 @@ column_info <- tribble(
 Now you can generate the funky heatmap:
 
 ```{r include=FALSE}
-g <- funky_heatmap(data, column_info = column_info, expand = list(xmax = 4))
+g <- funky_heatmap(data, column_info = column_info, position_args = position_arguments(expand_xmax = 4))
 knitr::opts_chunk$set(fig.width = g$width, fig.height = g$height)
 ```
 
 ```{r eval=FALSE}
-funky_heatmap(data, column_info = column_info, expand = list(xmax = 4))
+funky_heatmap(
+  data,
+  column_info = column_info,
+  position_args = position_arguments(expand_xmax = 4)
+)
 ```
 
 ```{r heatmap2, echo=FALSE}
