+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   executable: executable.html
   mtcars: mtcars.html
   scIB: scIB.html
-last_built: 2025-01-15T15:04Z
+last_built: 2025-01-27T09:55Z
 urls:
   reference: https://funkyheatmap.github.io/funkyheatmap/reference
   article: https://funkyheatmap.github.io/funkyheatmap/articles
