### 2025-01-28 5a73a93
Fix paper rendering
