### 2025-01-28 d8f855a
Update rendering figures
