### 2025-02-06 50e1e12
Update paper with editor comments. (#56)
* Quote affiliations

* Update affiliations

* Typo

* Add references

* Typo

* Update paper with reviewer comments

* Fix paper rendering

* Fix paper rendering

* Fix rendeering

* Fix rendering

* Update rendering figures

* Apply suggestions from code review

Co-authored-by: Robrecht Cannoodt <rcannood@gmail.com>

* Remove redundant figure

---------

Co-authored-by: Robrecht Cannoodt <rcannood@gmail.com>
