### 2025-02-06 bc1e332
Update JS (#57)
* fix references to js implementation

* update
