### 2025-02-06 #7698 @LouiseDck
Thanks @meenakshi-kushwaha and @rhenkin for agreeing to review!

In the meantime we've addressed comments made by @jromanowska in the [pre-review issue](https://github.com/openjournals/joss-reviews/issues/7590#issuecomment-2572510211). This has led to an updated version of the paper: we reworded a sentence and added a usage section.
We've also updated the README example, [as signalled by @rhenkin](https://github.com/funkyheatmap/funkyheatmap/issues/54)
