### 2025-02-28 #7698 @meenakshi-kushwaha
@jromanowska I have reviewed the paper and agree with @rhenkin's excellent assessment above, especially about comparing and explaining added functionality for their package over what existing packages like gt and complexheatmap offer, and a more gentle, step-by-step introduction using the mtcars example. Many of the checklist items are related to these points, so i have left them unchecked for now, and hope authors are able to address the concerns.
