### 2025-03-07 f5bf18c
remove unused footnotes
