### 2025-03-24 36fb5f6
Add gradual example
