+++ CONTRIBUTING.md
@@ -1,0 +1,25 @@
+# Contributing
+
+Thanks for considering contributing to funkyheatmap ❤️
+
+We welcome and encourage any contributions to this project, such as changes in documentation, bug fixes, extra functionality and bug reports.
+
+## I have a question / I found a bug
+Please ensure that you've read the [documentation](https://funkyheatmap.github.io/funkyheatmap/index.html).
+Before you ask questions or report new bugs, please search for existing [issues](https://github.com/funkyheatmap/funkyheatmap/issues).
+
+If the question or issue remains, then please open a [new issue](https://github.com/funkyheatmap/funkyheatmap/issues/new).
+Please provide context, and if possible, a minimal reproducable example.
+
+## I want to fix a bug or add functionality
+
+- Please submit a [new issue](https://github.com/funkyheatmap/funkyheatmap/issues/new), if no relevant issue already exists
+- Fork the repository on GitHub
+- Make your changes on your forked copy
+- Ensure that you add documentation and tests for your changes
+- Make sure that `rcmdcheck::rcmdcheck()` runs succesfully
+- Update NEWS.md with changes made
+- Submit a pull request and ensure that all github actions pass
+- Wait for the package maintainers to comment on the pull request
+
+
