### 2025-03-24 e503233
Update paper statement of need (#61)
* update paper

* Add citations

* remove unused footnotes
