+++ paper/paper.bib
@@ -130,6 +130,52 @@
   year={2024}
 }
 
+@article{complexheatmap1,
+  author = {Gu, Zuguang},
+  title = {Complex heatmap visualization},
+  journal = {iMeta},
+  volume = {1},
+  number = {3},
+  pages = {e43},
+  keywords = {bioconductor, clustering, complex heatmap, R package, visualization},
+  doi = {https://doi.org/10.1002/imt2.43},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/imt2.43},
+  eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/imt2.43},
+  abstract = {Abstract Heatmap is a widely used statistical visualization method on matrix-like data to reveal similar patterns shared by subsets of rows and columns. In the R programming language, there are many packages that make heatmaps. Among them, the ComplexHeatmap package provides the richest toolset for constructing highly customizable heatmaps. ComplexHeatmap can easily establish connections between multisource information by automatically concatenating and adjusting a list of heatmaps as well as complex annotations, which makes it widely applied in data analysis in many fields, especially in bioinformatics, to find hidden structures in the data. In this article, we give a comprehensive introduction to the current state of ComplexHeatmap, including its modular design, its rich functionalities, and its broad applications.},
+  year = {2022}
+}
+
+@article{complexheatmap2,
+  title = {Complex heatmaps reveal patterns and correlations in multidimensional genomic data},
+  author = {Zuguang Gu and Roland Eils and Matthias Schlesner},
+  journal = {Bioinformatics},
+  doi = {10.1093/bioinformatics/btw313},
+  year = {2016},
+}
+
+@Manual{superheat,
+  title = {superheat: A Graphical Tool for Exploring Complex Datasets Using Heatmaps},
+  author = {Rebecca Barter and Bin Yu},
+  year = {2017},
+  note = {R package version 0.1.0},
+  url = {https://CRAN.R-project.org/package=superheat},
+}
+
+@Manual{pheatmap,
+  title = {pheatmap: Pretty Heatmaps},
+  author = {Raivo Kolde},
+  year = {2019},
+  note = {R package version 1.0.12},
+  url = {https://CRAN.R-project.org/package=pheatmap},
+}
+
+@Manual{gt,
+  title = {gt: Easily Create Presentation-Ready Display Tables},
+  author = {Richard Iannone and Joe Cheng and Barret Schloerke and Ellis Hughes and Alexandra Lauer and JooYoung Seo and Ken Brevoort and Olivier Roy},
+  year = {2024},
+  note = {R package version 0.11.1},
+  url = {https://CRAN.R-project.org/package=gt},
+}
 
   ggplot2_wickham2009: 10.1007/978-0-387-98141-3
   matplotlib2dgraphics_hunter2007: 10.1109/MCSE.2007.55
