### 2025-03-28 f79af07
remove old mtcars vignette
