+++ vignettes/mtcars_gradual.Rmd
@@ -1,295 +1,0 @@
----
-title: "Demo with mtcars, complex"
-vignette: >
-  %\VignetteIndexEntry{Demo with mtcars}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteEngine{knitr::rmarkdown}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  dev = 'png',
-  fig.width=8, 
-  fig.height=11.5
-)
-```
-
-We will use the mtcars dataset to showcase how you can use **{funkyheatmap}** to visualise a dataset. We will gradually add features to 
-the plot to show how you can customise it to your liking.
-
-First, we load in the data. We move the rownames to a column named `id`, as that is required by **{funkyheatmap}** to display the names.
-Next, we sort the dataframe on the `mpg` column in descending order, and select the top 30 rows.
-
-```{r load-data}
-library(funkyheatmap)
-library(dplyr, warn.conflicts = FALSE)
-library(tibble, warn.conflicts = FALSE)
-library(purrr, warn.conflicts = FALSE)
-
-data("mtcars")
-
-data <- mtcars %>%
-  rownames_to_column("id") %>%
-  arrange(desc(mpg)) %>%
-  head(30)
-```
-
-We can plot this data frame without any additional formatting, though it doesn't look very nice:
-```{r basic-fh}
-funky_heatmap(data)
-```
-
-## Adding column_info
-
-We can add column information to the heatmap by specifying the `column_info` argument. 
-This argument should be a dataframe or a tibble, with each row corresponding to a column in the heatmap.
-The dataframe should have the following column:
-- `id (character)`: the name of the column in the heatmap
-
-Other columns are optional and will determine the appearance of the heatmap.
-If you want to group some of the columns together, usually because they are semantically related,
-or another logical grouping exists, we need to sepficy the `group` column in the `column_info` dataframe.
-
-```{r column-group}
-cinfo <- tibble(
-  id = colnames(data),
-  group = c(NA, "Overall", "Engine", "Engine", "Engine", "Transmission", "Overall", "Performance", "Engine", "Transmission", "Transmission", "Engine"),
-  options = lapply(seq(12), function(x) lst())
-)
-cinfo
-```
-
-```{r column-info-fh}
-funky_heatmap(data, column_info = cinfo)
-```
-
-We can see that this does not work: we need to manually rearrange the columns in the `column_info` dataframe
-so that the columns in the same column groups are adjacent to each other.
-
-```{r column-group-sort}
-data <- data[, c("id", "qsec", "mpg", "wt", "cyl", "carb", "disp", "hp", "vs", "drat", "am", "gear")]
-
-cinfo <- tibble(
-  id = colnames(data),
-  group = c(NA, "Performance", rep("Overall", 2), rep("Engine", 5), rep("Transmission", 3)),
-  options = lapply(seq(12), function(x) lst())
-)
-cinfo
-
-funky_heatmap(data, column_info = cinfo)
-```
-
-We should probably also add a `name` column, to make the column labels more informative.
-```{r column-info-name}
-cinfo$name <- c("Model", "1/4 mile time", "Miles per gallon", "Weight", "Number of cylinders", "Carburetors", "Displacement", "Horsepower", "Engine type", "Rear axle ratio", "Transmission", "Forward gears")
-funky_heatmap(data, column_info = cinfo)
-```
-
-This is better, we can at least read the column groups now. However, the column groups are not very clear: 
-we can add a `palette` column to the `column_info` dataframe to make the column groups more visually distinct.
-
-## Adding palettes
-
-Just adding a `palette` column is not good enough: we need to specify a separate palette list as well.
[…]
