+++ vignettes/mtcars.Rmd
@@ -1,27 +1,27 @@
 ---
-title: "Demo with mtcars"
+title: "Demo with mtcars, complex"
 vignette: >
   %\VignetteIndexEntry{Demo with mtcars}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
 ---
 
-
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  dev = 'png',
+  fig.width=8, 
+  fig.height=11.5
 )
 ```
 
-<!--
-rerun this to update tribble formatting:
-```r
-dynbenchmark::reformat_tribbles("vignettes/mtcars.Rmd")
-```
--->
+We will use the mtcars dataset to showcase how you can use **{funkyheatmap}** to visualise a dataset. We will gradually add features to 
+the plot to show how you can customise it to your liking.
 
-Load data
+First, we load in the data. We move the rownames to a column named `id`, as that is required by **{funkyheatmap}** to display the names.
+Next, we sort the dataframe on the `mpg` column in descending order, and select the top 30 rows.
+
 ```{r load-data}
 library(funkyheatmap)
 library(dplyr, warn.conflicts = FALSE)
@@ -33,115 +33,263 @@ data("mtcars")
 data <- mtcars %>%
   rownames_to_column("id") %>%
   arrange(desc(mpg)) %>%
-  head(20) %>%
-  mutate(type = c(rep("images/ice.png", 10), rep("images/electric.png", 10)))
+  head(30)
 ```
 
 We can plot this data frame without any additional formatting, though it doesn't look very nice:
-
-```{r}
-g <- funky_heatmap(data)
+```{r basic-fh}
+funky_heatmap(data)
 ```
 
-```{r include=FALSE}
-# set width and height for upcoming plot
-knitr::opts_chunk$set(fig.width = g$width, fig.height = g$height)
-```
+## Adding column_info
 
-```{r heatmap_preview}
-g
-```
+We can add column information to the heatmap by specifying the `column_info` argument. 
+This argument should be a dataframe or a tibble, with each row corresponding to a column in the heatmap.
+The dataframe should have the following column:
+- `id (character)`: the name of the column in the heatmap
 
+Other columns are optional and will determine the appearance of the heatmap.
+If you want to group some of the columns together, usually because they are semantically related,
+or another logical grouping exists, we need to sepficy the `group` column in the `column_info` dataframe.
 
-By defining a few additional formatting parameters, we can get the plot to look much nicer.
-
-## Column info 
-
-```{r column-info}
-column_info <- tribble(
-  ~id,     ~group,         ~name,                      ~geom,        ~palette,    ~options,
-  "id",    NA,             "",                         "text",       NA,          list(hjust = 0, width = 6),
-  "mpg",   "overall",      "Miles / gallon",           "bar",        "palette1",  list(width = 4, legend = FALSE),
-  "cyl",   "overall",      "Number of cylinders",      "bar",        "palette2",  list(width = 4, legend = FALSE),
-  "disp",  "group1",       "Displacement (cu.in.)",    "funkyrect",  "palette1",  lst(),
-  "hp",    "group1",       "Gross horsepower",         "funkyrect",  "palette1",  lst(),
-  "drat",  "group1",       "Rear axle ratio",          "funkyrect",  "palette1",  lst(),
-  "wt",    "group1",       "Weight (1000 lbs)",        "funkyrect",  "palette1",  lst(),
-  "qsec",  "group2",       "1/4 mile time",            "circle",     "palette2",  lst(),
-  "vs",    "group2",       "Engine",                   "circle",     "palette2",  lst(),
[…]
