+++ vignettes/mtcars.Rmd
@@ -1,7 +1,7 @@
 ---
-title: "Demo with mtcars, complex"
+title: "Functionality walkthrough (using the mtcars dataset)"
 vignette: >
-  %\VignetteIndexEntry{Demo with mtcars}
+  %\VignetteIndexEntry{Functionality walkthrough (using the mtcars dataset)}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
 ---
