### 2025-03-29 7c06f4d
Add gradual example to creating more complex plots (#62)
* Add gradual example

* remove old mtcars vignette

* Adapt title
