### 2025-03-31 b10ed41
Update paper fig2 size
