### 2025-03-31 19f9d45
update paper to display column info well
