### 2025-03-31 3691e83
Update render_figure
