+++ paper/paper.md
@@ -125,146 +125,7 @@ The table presents the suggested visualisation methods (geoms) based on the data
 In order to produce a `{funkyheatmap}` visualisation, you need to provide the data in the form of a dataframe, which also must contain a column named `id`.
 If you provide no other information, a basic visualisation will be provided, but customization is possible by provinding additional information, such as a `column_info` dataframe which details how the columns in the dataframe get translated into different geoms, or a `row_groups` dataframe which allows you to group rows in the visualisation.
 
-As an example, the following code produces \autoref{fig-mtcars}. For an in-depth explanation of how we come to 
-all these data structures and tribbles, please refer to the vignette named `mtcars`.
-
-```r
-library(funkyheatmap)
-library(dplyr, warn.conflicts = FALSE)
-library(tibble, warn.conflicts = FALSE)
-library(RColorBrewer)
-
-data("mtcars")
-data <- mtcars %>%
-  rownames_to_column("id") %>%
-  arrange(desc(mpg)) %>%
-  head(30)
-# sort the columns in logical groupings (see column_info group column)
-# sort the columns in logical groupings (see column_info group column)
-data <- data[, c("id", "qsec", "mpg", "wt", "cyl", "carb", "disp", 
-                 "hp", "vs", "drat", "am", "gear")]
-
-# change data to use images
-# change the am: if 0 go to "automatic", if 1 go to "manual"
-data[data$am == 0, "am"] <- "automatic"
-data[data$am == 1, "am"] <- "manual"
-# change the vs: if 0 go to "vengine", if 1 go to "straight"
-data[data$vs == 0, "vs"] <- "vengine"
-data[data$vs == 1, "vs"] <- "straight"
-
-column_info_part1 <- tribble(
-  ~id,     ~group,         ~name,                   ~geom,
-  "id",    "",             "Model",                 "text",
-  "qsec",  "Performance",  "1/4 mile time",         "bar",
-  "mpg",   "Overall",      "Number of cylinders",   "bar",
-  "wt",    "Overall",      "Weight (1000 lbs)",     "bar", 
-  "cyl",   "Engine",       "Number of cylinders",   "rect",  
-  "cyl",   "Engine",       "",                      "text",
-  "carb",  "Engine",       "Carburetors",           "rect",
-  "carb",  "Engine",       "",                      "text",    
-  "disp",  "Engine",       "Displacement",          "funkyrect",
-  "hp",    "Engine",       "Horsepower",            "funkyrect",
-  "vs",    "Engine",       "Engine type",           "image",
-  "drat",  "Transmission", "Rear axle ratio",       "funkyrect",
-  "am",    "Transmission", "Transmission",          "image",
-  "gear",  "Transmission", "# Forward gears",       "rect",
-  "gear",  "Transmission", "",                      "text"
-)
-
-column_info_part2 <- tribble(
-  ~palette,               ~options,
-  NA,                     lst(),
-  "perf_palette",         lst(),
-  "overall_palette",      lst(),
-  "overall_palette",      lst(),
-  "engine_palette",       lst(),
-  "black",                lst(overlay = TRUE),
-  "engine_palette",       lst(),
-  "black",                lst(overlay = TRUE),
-  "engine_palette",       lst(),
-  "engine_palette",       lst(),
-  "engine_palette",       lst(directory = "vignettes/images", extension = "png"),
-  "transmission_palette", lst(),
-  "transmission_palette", lst(directory = "vignettes/images", extension = "png"),
-  "transmission_palette", lst(),
[…]
