### 2025-04-01 e1bd872
Add better mtcars reference vignette (#63)
* Add better mtcars reference vignette

* Update paper fig2 size

* update paper to display column info well

* Update render_figure

* Remove badly rendered figure2

* Paper change
