+++ paper/paper.md
@@ -74,6 +74,7 @@ affiliations:
 
 engine: knitr
 bibliography: paper.bib
+code-block-font-size: \tiny
 ---
 
 *: Shared first authors, †: Shared last authors.
@@ -124,44 +125,9 @@ The table presents the suggested visualisation methods (geoms) based on the data
 In order to produce a `{funkyheatmap}` visualisation, you need to provide the data in the form of a dataframe, which also must contain a column named `id`.
 If you provide no other information, a basic visualisation will be provided, but customization is possible by provinding additional information, such as a `column_info` dataframe which details how the columns in the dataframe get translated into different geoms, or a `row_groups` dataframe which allows you to group rows in the visualisation.
 
-As an example, the following code produces \autoref{fig-mtcars}.
+As an example, \autoref{fig-mtcars} shows a visualisation of the mtcars dataset. For an in-depth explanation on how to fine-tune this figure and the different data structures involved, please see [`vignette("mtcars", "funkyheatmap")`](https://funkyheatmap.github.io/funkyheatmap/articles/mtcars.html).
 
-```r
-library(funkyheatmap)
-library(dplyr, warn.conflicts = FALSE)
-library(tibble, warn.conflicts = FALSE)
-
-data("mtcars")
-
-data <- mtcars %>%
-  rownames_to_column("id") %>%
-  arrange(desc(mpg))
-
-column_info <- tribble(
-  ~id,     ~group,         ~name,                      ~geom,        ~palette,  
-  "id",    "",             "",                         "text",       NA,        
-  "mpg",   "overall",      "Miles / gallon",           "bar",        "palette1",
-  "cyl",   "overall",      "Number of cylinders",      "bar",        "palette2", 
-  "disp",  "group1",       "Displacement (cu.in.)",    "funkyrect",  "palette1", 
-  "hp",    "group1",       "Gross horsepower",         "funkyrect",  "palette1", 
-  "drat",  "group1",       "Rear axle ratio",          "funkyrect",  "palette1", 
-  "wt",    "group1",       "Weight (1000 lbs)",        "funkyrect",  "palette1", 
-  "qsec",  "group2",       "1/4 mile time",            "circle",     "palette2", 
-  "vs",    "group2",       "Engine",                   "circle",     "palette2", 
-  "am",    "group2",       "Transmission",             "circle",     "palette2", 
-  "gear",  "group2",       "# Forward gears",          "circle",     "palette2", 
-  "carb",  "group2",       "# Carburetors",            "circle",     "palette2",
-)
-
-funky_heatmap(
-  data,
-  column_info = column_info,
-  position_args = position_arguments(expand_xmax = 4)
-)
-
-```
-
-![An example of a `{funkyheatmap}` visualisation using the mtcars dataset.\label{fig-mtcars}](figure2.svg){width="12cm"}
+![An example of a `{funkyheatmap}` visualisation using the mtcars dataset.\label{fig-mtcars}](figure2.pdf){width="15cm"}
 
 # Conclusion
 `{funkyheatmap}` streamlines the creation of publication-quality visualisation for mixed data types, empowering researchers and data scientists to communicate their results effectively. The development of `funkyheatmappy` (Python) and `funkyheatmapjs` (JavaScript) will further expand the accessibility and functionality of this visualisation solution.
