+++ paper/render_figure.R
@@ -1,6 +1,7 @@
 library(funkyheatmap)
 library(tidyverse)
 library(Cairo)
+library(RColorBrewer)
 
 legends <- list(
   list(title = "Bar", palette = "qc", enabled = FALSE, geom = "bar"),
@@ -38,34 +39,139 @@ df <- tibble::tribble(
 )
 knitr::kable(df)
 
+# Load the data, add the `id` column and sort based on mpg
 data("mtcars")
-
 data <- mtcars %>%
   rownames_to_column("id") %>%
-  arrange(desc(mpg))
+  arrange(desc(mpg)) %>%
+  head(30)
+# sort the columns in logical groupings (see column_info group column)
+data <- data[, c("id", "qsec", "mpg", "wt", "cyl", "carb", "disp", 
+                 "hp", "vs", "drat", "am", "gear")]
 
-column_info <- tribble(
-  ~id,     ~group,         ~name,                      ~geom,        ~palette,   
-  "id",    "",             "",                         "text",       NA,         
-  "mpg",   "overall",      "Miles / gallon",           "bar",        "palette1",
-  "cyl",   "overall",      "Number of cylinders",      "bar",        "palette2",
-  "disp",  "group1",       "Displacement (cu.in.)",    "funkyrect",  "palette1", 
-  "hp",    "group1",       "Gross horsepower",         "funkyrect",  "palette1", 
-  "drat",  "group1",       "Rear axle ratio",          "funkyrect",  "palette1", 
-  "wt",    "group1",       "Weight (1000 lbs)",        "funkyrect",  "palette1", 
-  "qsec",  "group2",       "1/4 mile time",            "circle",     "palette2", 
-  "vs",    "group2",       "Engine",                   "circle",     "palette2", 
-  "am",    "group2",       "Transmission",             "circle",     "palette2", 
-  "gear",  "group2",       "# Forward gears",          "circle",     "palette2", 
-  "carb",  "group2",       "# Carburetors",            "circle",     "palette2",
+# change data to use images
+# change the am: if 0 go to "automatic", if 1 go to "manual"
+data[data$am == 0, "am"] <- "automatic"
+data[data$am == 1, "am"] <- "manual"
+# change the vs: if 0 go to "vengine", if 1 go to "straight"
+data[data$vs == 0, "vs"] <- "vengine"
+data[data$vs == 1, "vs"] <- "straight"
+
+column_info_part1 <- tribble(
+  ~id,     ~group,         ~name,                   ~geom,
+  "id",    "",             "Model",                 "text",
+  "qsec",  "Performance",  "1/4 mile time",         "bar",
+  "mpg",   "Overall",      "Number of cylinders",   "bar",
+  "wt",    "Overall",      "Weight (1000 lbs)",     "bar", 
+  "cyl",   "Engine",       "Number of cylinders",   "rect",  
+  "cyl",   "Engine",       "",                      "text",
+  "carb",  "Engine",       "Carburetors",           "rect",
+  "carb",  "Engine",       "",                      "text",    
+  "disp",  "Engine",       "Displacement",          "funkyrect",
+  "hp",    "Engine",       "Horsepower",            "funkyrect",
+  "vs",    "Engine",       "Engine type",           "image",
+  "drat",  "Transmission", "Rear axle ratio",       "funkyrect",
+  "am",    "Transmission", "Transmission",          "image",
+  "gear",  "Transmission", "# Forward gears",       "rect",
+  "gear",  "Transmission", "",                      "text"
 )
+
+column_info_part2 <- tribble(
+  ~palette,               ~options,
+  NA,                    lst(),
[…]
