### 2025-04-01 #7698 @LouiseDck
Sorry for the delay!

> The repo/README currently lacks a contributing guideline section or document. You could add a CONTRIBUTING.md similar to [https://contributing.md/example/](https://contributing.md/example/)

We’ve added this [here](https://github.com/funkyheatmap/funkyheatmap/blob/main/CONTRIBUTING.md)

> Looking at the vignettes and examples, I think having a step-by-step example would be better than the current version. I think most people will start with the mtcars example,, and the first thing they will see is a tribble object and all parameters already defined with no explanation. I suggest having a more gentle introduction to the column_info par e.g. changing the shapes first, then palettes, etc. I'd also argue against using the tribble object, and to instead rely on structures that people may be more familiar with, e.g. vectors, data.frames (or tibbles) created from vectors and cbind too. Or possibly creating an initial column_info data.frame and then using the dplyr syntax e.g. mutate, case_when to change it step-by-step. I believe this could then be easily transferred to the example usage section of the paper too.

Thanks for this perspective, you’re right that the jump in complexity is too big. I’ve changed the mtcars vignette, which you can find [here](https://funkyheatmap.github.io/funkyheatmap/articles/mtcars.html). It starts with a basic visualisation of the mtcars dataset, and gradually turns it into a nicely visualised dataset by providing `column_info`, `column_groups`, `palettes` etc, and shows the result of all these steps visually as well.

We now refer to this vignette in the paper too.

I do think the tribble objects have value, as it’s a more intuitive way of seeing the data laid out for this kind of visualisation, so we haven’t changed anything in the more complex visualisations. I hope that the more gradual approach in the mtcars vignette helps here as well.

> I see how the package facilitates the creation of complex visual tables, but I'm not convinced the alternatives are ggplot or D3. Heatmap (e.g. ComplexHeatmap, pheatmap etc) and table-creating packages (e.g. gt) are more concrete alternatives that users may be more familiar with and would be more fitting entries in the statement of need section.

We've changed the statement of need in the paper to compare with pheatmap, superheat, ComplexHeatmap and gt as well.

> I also think there is a big jump from the mtcars example to the replications. I admit I don't have a good solution to this, but I'm wondering if it would be possible to have a more relatable example of usage with an actual benchmark computation? For example, it could be using base R functions to run different regression models or clustering on the mtcars data and then using funkyheatmap to summarise results.

I really appreciate the comment, but I do think this would be out of scope for this paper. The software doesn’t facilitate running a benchmark, nor does it try to guide the user on how to best do it.
I think good benchmarks require a lot of thought and experience and careful considerations of the problem you want to benchmark and the context in which this problem exists. Trying to show this in an artificial benchmark in a single vignette would be misleading as to the thought and time benchmarking takes.

However, I think the main point of this comment relates very much to one of the previous comments, that is, the jump in complexity. Here again, I hope that [the added vignette](https://funkyheatmap.github.io/funkyheatmap/articles/mtcars.html) helps in bridging that gap.

Thanks @rhenkin and @meenakshi-kushwaha for the comments, I hope we've addressed them sufficiently here.
