### 2025-04-02 #7698 @rhenkin
Hi all,

Thanks @LouiseDck I think the mtcars vignette looks great and it's a much gentler starting point. I think the progressive additions to the tibble make more sense to reader, I felt that in the complex examples the very large tibbles were a lot to process.

I'm happy with all the changes and completed the review.

Rafael
