### 2025-04-02 #7698 @jromanowska
@rhenkin - please check off the last item on your list (Performance) if you're satisfied with the changes.
