### 2025-04-08 #7698 @jromanowska
In addition, you have quite a long list of co-authors. Please, add a section to the paper explaining contributions: who wrote the code or maybe reviewed the code, etc. Please remember that purely organizational or financial roles are not enough to be a co-author - see [the documentation on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship).
