+++ paper/paper.bib
@@ -138,7 +138,7 @@
   number = {3},
   pages = {e43},
   keywords = {bioconductor, clustering, complex heatmap, R package, visualization},
-  doi = {https://doi.org/10.1002/imt2.43},
+  doi = {10.1002/imt2.43},
   url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/imt2.43},
   eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/imt2.43},
   abstract = {Abstract Heatmap is a widely used statistical visualization method on matrix-like data to reveal similar patterns shared by subsets of rows and columns. In the R programming language, there are many packages that make heatmaps. Among them, the ComplexHeatmap package provides the richest toolset for constructing highly customizable heatmaps. ComplexHeatmap can easily establish connections between multisource information by automatically concatenating and adjusting a list of heatmaps as well as complex annotations, which makes it widely applied in data analysis in many fields, especially in bioinformatics, to find hidden structures in the data. In this article, we give a comprehensive introduction to the current state of ComplexHeatmap, including its modular design, its rich functionalities, and its broad applications.},
@@ -159,6 +159,7 @@
   year = {2017},
   note = {R package version 0.1.0},
   url = {https://CRAN.R-project.org/package=superheat},
+  doi = {10.32614/cran.package.superheat},
 }
 
 @Manual{pheatmap,
@@ -167,6 +168,7 @@
   year = {2019},
   note = {R package version 1.0.12},
   url = {https://CRAN.R-project.org/package=pheatmap},
+  doi = {10.32614/cran.package.pheatmap},
 }
 
 @Manual{gt,
@@ -175,6 +177,7 @@
   year = {2024},
   note = {R package version 0.11.1},
   url = {https://CRAN.R-project.org/package=gt},
+  doi = {10.32614/CRAN.package.gt},
 }
 
   ggplot2_wickham2009: 10.1007/978-0-387-98141-3
