### 2025-04-08 fffc92e
Add author contributions (#65)
