+++ paper/paper.md
@@ -125,7 +125,7 @@ The table presents the suggested visualisation methods (geoms) based on the data
 In order to produce a `{funkyheatmap}` visualisation, you need to provide the data in the form of a dataframe, which also must contain a column named `id`.
 If you provide no other information, a basic visualisation will be provided, but customization is possible by provinding additional information, such as a `column_info` dataframe which details how the columns in the dataframe get translated into different geoms, or a `row_groups` dataframe which allows you to group rows in the visualisation.
 
-As an example, \autoref{fig-mtcars} shows a visualisation of the mtcars dataset. For an in-depth explanation on how to fine-tune this figure and the different data structures involved, please see the ['Getting started'](https://funkyheatmap.github.io/funkyheatmap/articles/funkyheatmap.html) article in the package documentation.
+As an example, \autoref{fig-mtcars} shows a visualisation of the mtcars dataset. For an in-depth explanation on how to fine-tune this figure and the different data structures involved, please see the [Getting started](https://funkyheatmap.github.io/funkyheatmap/articles/funkyheatmap.html) article in the package documentation.
 
 ![An example of a `{funkyheatmap}` visualisation using the mtcars dataset.\label{fig-mtcars}](figure2.pdf){width="15cm"}
 
