### 2025-04-09 96769db
Update mtcars vignette for CRAN (#67)
* update mtcars vignette for CRAN

* update reference in paper

* update references to vignette

* fix vignette

* Color warning red

---------

Co-authored-by: LouiseDck <louise.deconinck@gmail.com>
