+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   executable: executable.html
   funkyheatmap: funkyheatmap.html
   scIB: scIB.html
-last_built: 2025-04-09T08:45Z
+last_built: 2025-04-09T12:39Z
 urls:
   reference: https://funkyheatmap.github.io/funkyheatmap/reference
   article: https://funkyheatmap.github.io/funkyheatmap/articles
