### 2025-04-09 #7698 @LouiseDck
- I've added a section detailing the contributions of all authors.
- I've double checked the author affiliations and DOIs
- The version number of the release is [0.5.2](https://github.com/funkyheatmap/funkyheatmap/releases/tag/0.5.2)
- The release is archived on Zenodo with doi: 10.5281/zenodo.15182434
- The title, author list and license match the ones in the paper and on the software
