### 2025-04-09 #7698 @jromanowska
Thank you for doing these tasks so fast!

In the "Author contribution" section, in the paper, you write:

> "A.C. authored the Python version of the package. N.M. authored the Javascript version of the package."

This submission, however, is only about the R version - could you explain why these authors are included? Have they contributed to the R package also?
