### 2025-04-10 3f5d884
Correct affiliations and author contributions (#68)
* Reorder and fix affiliations

* Add correct author contributions
