### 2025-04-10 #7698 @LouiseDck
- The later versions of the three packages were designed and developed together, so quite some functionality was decided on together. I've added this line in the Author Contribution sector to reflect this:

> R.C., L.D., L.Z., A.C. and N.M. participated in design meetings and discussions that shaped the software design and functionality of all packages.

- I've reordered the affiliations, checked all of them and affiliation no.8 was indeed wrongly not assigned. Thanks for catching that!
- I've updated the affiliations on Zenodo, and they should now be identical to the ones on the paper.

Do you want me to publish a new release because of these changes in the paper?
