### 2025-04-10 #7698 @jromanowska
One more thing about the paper. In the Conclusions, you write:

> "The development of funkyheatmappy (Python) and funkyheatmapjs (JavaScript) will further expand the accessibility and functionality of this visualization solution."

I don't quite understand this sentence. Do you mean that the fact that you have three packages, implemented in three different languages, makes the method more accessible? Please reword.
