### 2025-04-10 4d6d7a8
Small corrections to  the paper text
