### 2025-04-10 7b5ac7d
Reword other languages section (#70)
