### 2025-04-10 #7698 @LouiseDck
> One more thing about the paper. In the Conclusions, you write:
>
> > "The development of funkyheatmappy (Python) and funkyheatmapjs (JavaScript) will further expand the accessibility and functionality of this visualization solution."
>
> I don't quite understand this sentence. Do you mean that the fact that you have three packages, implemented in three different languages, makes the method more accessible? Please reword.

We've reworded this as:

> The package has been ported to other programming languages, as `funkyheatmappy` (Python) and `funkyheatmapjs` (JavaScript). This expands the accessibility of this type of visualisation to non-R users. The JavaScript port incorporates interactive elements, expanding on the visualization's functionality.

> I've also created a pull request with some suggestions of corrections.

Thanks for the suggestions, this got merged as well.
