### 2025-04-10 #7698 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
