### 2025-04-10 #7698 @danielskatz
👋 @LouiseDck - I've suggested two minor changes in https://github.com/funkyheatmap/funkyheatmap/pull/71 - please merge this, or let me know what you disagree with, then we can continue.
