### 2025-04-11 d66dd6d
Minor changes in paper.md (#71)
