### 2025-04-11 #7698 @rcannood
Thanks @danielskatz ! The suggested changes LGTM, so I merged them.
