+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   executable: executable.html
   funkyheatmap: funkyheatmap.html
   scIB: scIB.html
-last_built: 2025-04-10T15:17Z
+last_built: 2025-04-11T06:37Z
 urls:
   reference: https://funkyheatmap.github.io/funkyheatmap/reference
   article: https://funkyheatmap.github.io/funkyheatmap/articles
