### 2025-04-11 #7698 @danielskatz
Congratulations to  @LouiseDck (Louise Deconinck) and co-authors on your JOSS publication!!
